Add files via upload
Initial approach to the problem y dividing each of the 64 blocks further into 16 blocks and applying filters to derive different features, performing clustering and taking the majority cluster to be the class of the block. Problems - context of between blocks or between sub blocks is not included, method used to differentiate the two clusters is not well defined.